README: Added hardware and wiring plan - display is now 128x64 matrix
The character LCD was too limiting - a multi-line font could not be made to look
good enough with the extremely limited custom character support (8 is not many).
A new display has been acquired which is wired very similarly to the character
LCD, but is a graphical matrix display. This will require some changes to wiring
and a lot of changes to code - most likely a new Python/asyncio driver.